+++ MagicalLifeAPI/Load/ItemLoader.cs
@@ -11,7 +11,7 @@ namespace MagicalLifeAPI.Load
         {
             ItemRegistry.Initialize(new List<Type>()
             {
-                typeof(StoneChunk)
+                typeof(StoneRubble)
             });
         }
     }