+++ MagicalLifeAPI/DataTypes/Point2D.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using ProtoBuf;
 using System;
+using System.Globalization;
 
 namespace MagicalLifeAPI.DataTypes
 {
@@ -72,7 +73,7 @@ namespace MagicalLifeAPI.DataTypes
 
         public override string ToString()
         {
-            return "{ " + this.X.ToString() + ", " + this.Y.ToString() + " }";
+            return "{ " + this.X.ToString(CultureInfo.InvariantCulture) + ", " + this.Y.ToString(CultureInfo.InvariantCulture) + " }";
         }
     }
 }