+++ MagicalLifeAPI/Entity/Movement/EntityWorldMovement.cs
@@ -27,9 +27,6 @@ namespace MagicalLifeAPI.Entity.Movement
         /// <param name="entity"></param>
         public static void MoveEntity(Living entity)
         {
-            //Log.Debug("Entity movement speed: " + entity.Movement.GetValue().ToString());
-            //Log.Debug("Entity starting position: " + entity.ScreenLocation.ToString());
-
             ProtoQueue<PathLink> path = entity.QueuedMovement;
 
             while (entity.Movement.GetValue() > 0 && path.Count > 0)
@@ -39,12 +36,7 @@ namespace MagicalLifeAPI.Entity.Movement
                 Tile sourceTile = World.Data.World.Dimensions[entity.Dimension][section.Origin.X, section.Origin.Y];
                 Tile destinationTile = World.Data.World.Dimensions[entity.Dimension][section.Destination.X, section.Destination.Y];
                 Move(entity, sourceTile, destinationTile);
-
-                //Log.Debug("Entity position: " + entity.ScreenLocation.ToString());
             }
-
-            //Log.Debug("Entity end position: " + entity.ScreenLocation.ToString());
-            //Log.Debug("Entity movement speed after: " + entity.Movement.GetValue().ToString());
         }
 
         /// <summary>