+++ MagicalLifeGUIWindows/GUI/Join Game Menu/Input Boxes/PortInputBox.cs
@@ -6,7 +6,10 @@ namespace MagicalLifeGUIWindows.GUI.Join
 {
     public class PortInputBox : MonoInputBox
     {
-        public PortInputBox(bool isLocked) : base(TextureLoader.GUIInputBox100x50, TextureLoader.GUICursorCarrot, GetInitialLocation(), int.MaxValue, TextureLoader.FontMainMenuFont12x, isLocked, Rendering.Text.SimpleTextRenderer.Alignment.Left, true)
+        public PortInputBox(bool isLocked)
+            : base(TextureLoader.GUIInputBox100x50, TextureLoader.GUICursorCarrot, GetInitialLocation(),
+                  int.MaxValue, TextureLoader.FontMainMenuFont12x, isLocked,
+                  Rendering.Text.SimpleTextRenderer.Alignment.Left, true)
         {
         }